safari-600.7-branch/LayoutTests/http/tests/security/cross-frame-access-frameelement-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8080" from accessing a frame with origin "http://127.0.0.1:8000". Protocols, domains, and ports must match.
This test checks if frameElement is accessible from the same or cross origin iframe.

 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
[object HTMLIFrameElement]

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
null
